+++ src/lib/index.ts
@@ -5,5 +5,7 @@ export { extractDocTests, runDocTests } from "./analyzer/doctest.js";
 export { generateJSON } from "./generator/json.js";
 export { generateMarkdown } from "./generator/markdown.js";
 export { loadConfig, mergeConfig } from "./config/index.js";
+export { runPlugins } from "./plugins/index.js";
 export type { OxdocConfig } from "./config/index.js";
 export type * from "../types/index.js";
+export type * from "../types/plugin.js";